Fix CI

+++ src/generator_gemm_amx.c
@@ -376,7 +376,7 @@ LIBXSMM_API_INTERN void libxsmm_generator_gemm_decompress_KxM_mxfp4_tensor( libx
             LIBXSMM_X86_INSTR_PREFETCHT0,
             i_gp_scf,
             LIBXSMM_X86_GP_REG_UNDEF, 0,
-            (int)((im * l_vlen + ik * i_ldi) * i_micro_kernel_config->datatype_size_in + (long long)(i_xgemm_desc->c1/16)) );
+            (int)(((long long)(im * l_vlen + ik * i_ldi)) * i_micro_kernel_config->datatype_size_in + (long long)(i_xgemm_desc->c1/16)) );
       }
       libxsmm_x86_instruction_vec_compute_2reg( io_generated_code, LIBXSMM_X86_INSTR_VPMOVZXBW, i_micro_kernel_config->vector_name, l_scf_vreg, l_scf_vreg);
       libxsmm_x86_instruction_vec_compute_3reg_imm8( io_generated_code, LIBXSMM_X86_INSTR_VPCMPW, i_micro_kernel_config->vector_name, i_micro_kernel_config->zero_reg, l_scf_vreg, l_mask_expon, 0 );
